Move propagation.go to package .../levee/propagation.
